- Made all getters nullable in XFileSource - Added lazy-loading of a blob for the file content which fixes saveTo with html.dart when using a custom source - Fixed incorrect constructor in interface.dart - Moved TextXFileSource to common.dart file
